requirement.txt
ctransformers==0.2.27
sentence-transformers==0.3.0
pinecone  # ✅ Replaces deprecated `pinecone-client`
langchain>=0.1.0  # ✅ Updated to new structure
langchain-community  # ✅ Added for integrations like CTransformers
flask
pypdf
python-dotenv
-e .
